cvczxvxcvxv

+++ TelemetryDependencies/Models/Message.cs
@@ -8,8 +8,8 @@ namespace TelemetryDependencies.Models
     {
         public int Id { get; set; }
         public string Prefix { get; set; }
-        public byte MessageId { get; set; }
-        public byte Length { get; set; }
+        public int MessageId { get; set; }
+        public int Length { get; set; }
         public string Text { get; set; }
         public DateTime DateTime { get; set; }
     }
